Refactor BackCommand and introduce BackListener
Extract event handling from BackCommand to new BackListener class for better separation of concerns. This change improves code maintainability by decoupling command execution logic from event handling logic.
